Fixes integer overflow.

+++ src/dedispersion.cpp
@@ -85,8 +85,8 @@ void compute_partial_dedispersion(Images& images, int start_freq_idx, int freq_b
         return (window_start_idx + offset) % table_size;
     };
 
-    int cell_size = n_dms * table_size;
-    int width = cell_size * images.side_size;
+    unsigned long long cell_size = n_dms * table_size;
+    unsigned long long width = cell_size * images.side_size;
     
 
     int end_freq = start_freq_idx - freq_batch_size + 1;
@@ -141,8 +141,8 @@ void compute_partial_dedispersion(Images& images, int start_freq_idx, int freq_b
 }
 
 void clear_buffer(float* dm_starttime, int side_size, int n_dms, int start_idx, int table_size, int buffer_size){
-    int width = n_dms * table_size * side_size;
-    int cell_size = n_dms * table_size;
+    unsigned long long  width = n_dms * table_size * side_size;
+    unsigned long long  cell_size = n_dms * table_size;
     #pragma omp parallel for collapse(2)
     for(int x = 0; x < side_size; x++){
         for(int y = 0; y < side_size; y++){
@@ -159,8 +159,8 @@ void clear_buffer(float* dm_starttime, int side_size, int n_dms, int start_idx, 
 void get_elements(float *dm_starttime, int side_size, int n_dms, int dm_idx, int window_start_idx, int buffer_size, 
         int table_size, float norm_factor, int x, int y){
     
-    int width = n_dms * table_size * side_size;
-    int cell_size = n_dms * table_size;
+    unsigned long long  width = n_dms * table_size * side_size;
+    unsigned long long  cell_size = n_dms * table_size;
 
     for(int i {0}; i < buffer_size; i++){
         int idx = (window_start_idx + i) % table_size;
@@ -170,8 +170,8 @@ void get_elements(float *dm_starttime, int side_size, int n_dms, int dm_idx, int
 
 void dump_buffer(float *dm_starttime, int side_size, int n_dms, int window_start_idx, int buffer_size, 
         int table_size, std::string filename, int x_0, int y_0){
-    int width = n_dms * table_size * side_size;
-    int cell_size = n_dms * table_size;
+    unsigned long long width = n_dms * table_size * side_size;
+    unsigned long long cell_size = n_dms * table_size;
     std::ofstream out_file(filename, std::ios::app | std::ios::binary);
     #define write_to_file(X) out_file.write(reinterpret_cast<char*>(&X), sizeof(X))
